fix: split PipelineViewingOptions into small ones.
